requirements.txt
langgraph-sdk>=0.2.9
langgraph-checkpoint-sqlite>=3.0.0
langchain-community>=0.4.1
notebook
lxml
scikit-learn
pandas
pyarrow

langgraph>=1.0.3
langchain>=1.0.8
langgraph-cli[inmem]
python-dotenv
langchain-openai>=1.0.0
langsmith>=0.4.0